% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/an0_retrieve_species_info.R
\name{an0_retrieve_species_info}
\alias{an0_retrieve_species_info}
\title{Retrieve taxonomic information.}
\usage{
an0_retrieve_species_info(sciNames, verbose = FALSE)
}
\arguments{
\item{sciNames}{Vector of scientific names of species.}

\item{verbose}{Logical. Should extra information be reported on progress?}
}
\value{
A data frame, each row contains the information for one species, NA #' are added for pieces of information missing.
}
\description{
Assign a unique identifiers to species and build a data frame including
their taxonomic hierarchy. Note that the data frame for the full set of
species is included in \code{fishList} (see \code{?fishList}) abd the code of thi
function shows how we retrieved the data.
}
\examples{
# Not run
# an0_retrieve_species_info(sps)

}
