% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/an1_gain_loss.R
\name{an1_gain_loss}
\alias{an1_gain_loss}
\title{Compute gains and losses}
\usage{
an1_gain_loss(x, nocc = 10, min_year_AHI = 1965, max_year_AHI = 1982,
  rm_outlier = TRUE, verbose = FALSE)
}
\arguments{
\item{x}{data set of observation see \code{?df_bsm_ahi}.}

\item{nocc}{Only species that have more than \code{nocc} occurrence are considered .}

\item{min_year_AHI}{First year for AHI (see \code{details} below).}

\item{max_year_AHI}{Last year for AHI (see \code{details} below).}

\item{rm_outlier}{a logical. Should outlier be removed? Default is set to \code{TRUE}.}

\item{verbose}{a logical. Should extra information be reported on progress? Default is set to \code{TRUE}.}
}
\value{
A list of six elements:
\itemize{
\item \code{sf_bsm_ahi}: a sf object including filtered data,
\item \code{gain_loss}: gains and losses for all lakes,
\item \code{beta_AHI}: beta diversity measure in AHI (historical survey),
\item \code{beta_BSM}: beta diversity measure in BSM (contempory survey),
\item \code{nm_ahi}: column names of ahi presence/absence columns in AHI,
\item \code{nm_bsm}: column names of bsm presence/absence columns in BSM.
}
}
\description{
Filter data and compute gains and losses
}
\details{
\code{min_year_AHI} and \code{max_year_AHI} are used to reproduce the analysis with
a subset of the dataset as used in Fig. S3.
}
\examples{
res0 <- an1_gain_loss(sf_bsm_ahi)
}
