% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/an2_occurrence_species.R
\name{an2_occurrence_species}
\alias{an2_occurrence_species}
\title{Occurrence analysis}
\usage{
an2_occurrence_species(x)
}
\arguments{
\item{x}{output of \code{\link[=an1_gain_loss]{an1_gain_loss()}}.}
}
\value{
A data frame with info about species
}
\description{
Occurrence analysis for individual species.
}
