% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/an4_contributions_beta.R
\name{an4_contributions_beta}
\alias{an4_contributions_beta}
\title{Contribution to beta diversity metrics}
\usage{
an4_contributions_beta(x, y)
}
\arguments{
\item{x}{output of \code{\link[=an1_gain_loss]{an1_gain_loss()}}.}

\item{y}{output of \code{\link[=an2_occurrence_species]{an2_occurrence_species()}}.}
}
\value{
A list of five elements:
\itemize{
\item \code{contrb}: contribution to beta diversity metrics;
\item \code{fsh_ahi}: number of baitfish species in AHI for every lake;
\item \code{fsh_bsm}: number of baitfish species in BSM for every lake;
\item \code{nofsh_ahi}: number of baitfish species in AHI for every lake;
\item \code{nofsh_bsm}: number of baitfish species in BSM for every lake.
}
}
\description{
Contribution to beta (spatial and temporal) of all species.
}
\examples{
res0 <- an1_gain_loss(sf_bsm_ahi)
res1 <- an2_occurrence_species(res0)
res <- an4_contributions_beta(res0, res1)
}
