% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_fished}
\alias{df_fished}
\title{Information about commercial aspects of fish species}
\format{A data frame of 8 columns and 39 rows.}
\usage{
df_fished
}
\description{
Information about commercial aspects of fish species
}
\keyword{datasets}
