% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_species_info}
\alias{df_species_info}
\title{Characteristics of Ontarian freshwater fish species}
\format{A data frame of 34 columns and 169 rows.}
\usage{
df_species_info
}
\description{
Characteristics of freshwater fish species included in the study retrieved
from the literature as explained in Cazelles (2019).
}
\references{
Cazelles (2019)
}
\author{
Kevin Cazelles
}
\keyword{datasets}
