% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig2_homogenization.R
\name{fig2_homogenization}
\alias{fig2_homogenization}
\title{Draw Figure 2}
\usage{
fig2_homogenization(x, cex_bas = 1, add_let = TRUE, addtitle = FALSE,
  asfinal = TRUE)
}
\arguments{
\item{x}{output of \code{\link[=an1_gain_loss]{an1_gain_loss()}}.}

\item{cex_bas}{magnification coefficient, controls the size of plot elements.}

\item{add_let}{logical. Should a letter for the panel be added? Default is set to \code{TRUE}.}

\item{addtitle}{a logical. Should a title be added. Default is set to \code{FALSE}?}

\item{asfinal}{logical. Should the figure be formatted for publication and save as a separate file? Note that in this case \code{cex_bas} is ignored.}
}
\description{
Perform rarefaction curves, test for homogenization and draw Figure 2 - homogenization and rarefaction curves.
}
\examples{
fig2_homogenization(an1_gain_loss(sf_bsm_ahi))
}
