% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig3_regressions.R
\name{fig3_regressions}
\alias{fig3_regressions}
\title{Create figure 2 - Regressions}
\usage{
fig3_regressions(x, mods, cex_bas = 1, asfinal = TRUE)
}
\arguments{
\item{x}{output of \code{\link[=an1_gain_loss]{an1_gain_loss()}}.}

\item{mods}{output of \code{\link[=an3_regressions]{an3_regressions()}}.}

\item{cex_bas}{magnification coefficient, controls the size of plot elements.}

\item{asfinal}{logical. Should the figure be formatted for publication and save as a separate file? Note that in this case \code{cex_bas} is ignored.}
}
\description{
Create figure 2 - Regressions
}
\examples{
res0 <- an1_gain_loss(sf_bsm_ahi)
res1 <- an2_occurrence_species(res0)
mods <- an3_regressions(res0, res1)
fig3_regressions(res0, mods)
}
