% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig4_table.R
\name{fig4_table}
\alias{fig4_table}
\title{Create figure 3}
\usage{
fig4_table(x, nlk, cex_bas = 5.4)
}
\arguments{
\item{x}{output of \code{\link[=an2_occurrence_species]{an2_occurrence_species()}}.}

\item{nlk}{number of lakes.}

\item{cex_bas}{magnification coefficient, controls the size of plot elements.}
}
\description{
Create Fig. 3 - Change in occupancy
}
\examples{
# res0 <- an1_gain_loss(sf_bsm_ahi)
# res1 <- an2_occurrence_species(res0)
# fig4_table(res1, nrow(res0[[1]]))
}
\references{
The image are available on line:
\itemize{
\item Hook \url{https://www.shareicon.net/miscellaneous-fishing-hook-sport-tools-and-utensils-steel-788675}
\item Lake Trout \url{http://phylopic.org/image/7f2cbb42-12b1-4481-8ac0-705eb7363c74/}
\item White sucker \url{http://phylopic.org/image/c657b67e-db70-4f8d-a66f-e6b287b148a6/}
\item Yellow perch \url{http://phylopic.org/image/5e431267-dc57-435d-a64f-b91d4c569677/}
}
}
