% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig5_piecharts.R
\name{fig5_piecharts}
\alias{fig5_piecharts}
\title{Draw Figure 5}
\usage{
fig5_piecharts(x, nlk, cex_bas = 3.3)
}
\arguments{
\item{x}{output of \code{\link[=an2_occurrence_species]{an2_occurrence_species()}}.}

\item{nlk}{number of lakes.}

\item{cex_bas}{magnification coefficient, controls the size of plot elements.}
}
\description{
Compute occupancy changes for every grouping and draw Figure 5.
}
\examples{
res0 <- an1_gain_loss(sf_bsm_ahi)
res1 <- an2_occurrence_species(res0)
}
