% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig6_contributions_beta.R
\name{fig6_contributions_beta}
\alias{fig6_contributions_beta}
\title{Draw Figure 6}
\usage{
fig6_contributions_beta(x, cex_bas = 6)
}
\arguments{
\item{x}{output of \code{\link[=an4_contributions_beta]{an4_contributions_beta()}}.}

\item{cex_bas}{magnification coefficient, controls the size of plot elements.}
}
\description{
Use the output of \code{\link[=an4_contributions_beta]{an4_contributions_beta()}} and draw Figure 6.
}
\examples{
res0 <- an1_gain_loss(sf_bsm_ahi)
res1 <- an2_occurrence_species(res0)
res2 <- an4_contributions_beta(res0, res1)
fig6_contributions_beta(res2)
}
