% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figS1_outliers.R
\name{figS1_outliers}
\alias{figS1_outliers}
\title{Draw Figure S1}
\usage{
figS1_outliers(x, cex_bas = 5)
}
\arguments{
\item{x}{output of \code{\link[=an1_gain_loss]{an1_gain_loss()}}.}

\item{cex_bas}{magnification coefficient, controls the size of plot elements.}
}
\description{
Draw Figure S1
}
