% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figS3_lossgain.R
\name{figS3_lossgain}
\alias{figS3_lossgain}
\title{Draw Figure S3}
\usage{
figS3_lossgain(x, cex_bas = 6)
}
\arguments{
\item{x}{output of \code{\link[=an1_gain_loss]{an1_gain_loss()}}.}

\item{cex_bas}{magnification coefficient, controls the size of plot elements.}
}
\description{
Draw Figure S3
}
