% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figS5_saturation.R
\name{figS5_saturation}
\alias{figS5_saturation}
\title{Draw Figure S5}
\usage{
figS5_saturation(x, cex_bas = 6)
}
\arguments{
\item{x}{output of \code{\link[=an1_gain_loss]{an1_gain_loss()}}.}

\item{cex_bas}{magnification coefficient, controls the size of plot elements.}
}
\description{
Perform model II linear regression and drawFigure S5.
}
