% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figS6_northward.R
\name{figS6_northward}
\alias{figS6_northward}
\title{Draw Figure S6}
\usage{
figS6_northward(x, cex_bas = 4.5)
}
\arguments{
\item{x}{output of \code{\link[=an2_occurrence_species]{an2_occurrence_species()}}.}

\item{cex_bas}{magnification coefficient, controls the size of plot elements.}
}
\description{
Draw Figure S6
}
