% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figS7_occupancy_trends.R
\name{figS7_occupancy_trends}
\alias{figS7_occupancy_trends}
\title{Draw Figure S7}
\usage{
figS7_occupancy_trends(x, cex_bas = 5)
}
\arguments{
\item{x}{output of \code{\link[=an2_occurrence_species]{an2_occurrence_species()}}.}

\item{cex_bas}{magnification coefficient, controls the size of plot elements.}
}
\description{
Draw Figure S7
}
