% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figS8_boats.R
\name{figS8_boats}
\alias{figS8_boats}
\title{Draw Figure S8}
\usage{
figS8_boats(x, cex_bas = 5)
}
\arguments{
\item{x}{output of \code{\link[=an1_gain_loss]{an1_gain_loss()}}.}

\item{cex_bas}{magnification coefficient, controls the size of plot elements.}
}
\description{
Draw Figure S8
}
