% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figZ1_fnf_regressions.R
\name{figZ1_fnf_regressions}
\alias{figZ1_fnf_regressions}
\title{Draw Figure Z1}
\usage{
figZ1_fnf_regressions(x, y, cex_bas = 6)
}
\arguments{
\item{x}{output of \code{\link[=an1_gain_loss]{an1_gain_loss()}}.}

\item{y}{output of \code{\link[=an4_contributions_beta]{an4_contributions_beta()}}.}

\item{cex_bas}{magnification coefficient, controls the size of plot elements.}
}
\description{
Draw Figure Z1
}
\details{
Figure Z1 was found in a previous version of the SI.
}
