% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figZ2_coldest.R
\name{figZ2_coldest}
\alias{figZ2_coldest}
\title{Draw Figure Z2}
\usage{
figZ2_coldest(x, y, cex_bas = 4)
}
\arguments{
\item{x}{output of \code{\link[=an1_gain_loss]{an1_gain_loss()}}.}

\item{y}{output of \code{\link[=an4_contributions_beta]{an4_contributions_beta()}}.}

\item{cex_bas}{magnification coefficient, controls the size of plot elements.}
}
\description{
Draw Figure Z2
}
\details{
Figure Z2 was found in a previous version of the SI.
}
