% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gadm_ontario}
\alias{gadm_ontario}
\title{Ontario boundaries}
\format{A \code{sf} object including the border of Ontario.}
\usage{
gadm_ontario
}
\description{
Ontario boundaries etrieved from https://gadm.org/ using \code{\link[raster:getData]{raster::getData()}} and convert to a \code{sf} object.
}
\keyword{datasets}
