% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{pipeline}
\alias{pipeline}
\title{Pipeline}
\usage{
pipeline(tnr = TRUE)
}
\arguments{
\item{tnr}{a logical. Should Times New Roman be used as a typeface?}
}
\description{
This function performs the analysis and creates all figures included
in Cazelles (2019).
}
