% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sf_bsm_ahi}
\alias{sf_bsm_ahi}
\title{Lake data for AHI and BSM survey}
\format{A \code{sf} object}
\usage{
sf_bsm_ahi
}
\description{
Lake data including species and presence absence and abiotic variables
both climatic variables and lake descriptors.
}
\references{
\itemize{
\item Dodge, D. P., et a. (1987). Manual of instructions: Aquatic habitat inventory surveys. Ontario Ministry of Natural Resources.
\item Sandstrom, S., Rawson, M., & Lester, N. P. (2013). Manual of instructions for broad-scale fish community monitoring using North American (NA1) and Ontario small mesh (ON2) gillnets. Ontario Ministry of Natural Resources.
\item McKenney, D. et al. (2013). The Forestry Chronicle. DOI: 10.5558/tfc2013-118.
}
}
\keyword{datasets}
